% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_soil_trapezoidal.R
\name{calc_soil_trapezoidal}
\alias{calc_soil_trapezoidal}
\title{Calculate weighted average of soil column based on trapezoidal rule}
\usage{
calc_soil_trapezoidal(
  depth1 = 0,
  depth2 = 5,
  depth3 = 15,
  depth4 = 30,
  value1,
  value2,
  value3,
  value4
)
}
\arguments{
\item{depth1}{Numeric. Depth of top soil column interval. Defaults to 0 cm.}

\item{depth2}{Numeric. Depth of second from top soil column interval. Defaults to 5 cm.}

\item{depth3}{Numeric. Depth of third from top soil column interval. Defaults to 15 cm.}

\item{depth4}{Numeric. Depth of bottom soil column interval. Defaults to 30 cm.}

\item{value1}{Numeric. Value of the soil property at depth1.}

\item{value2}{Numeric. Value of the soil property at depth2.}

\item{value3}{Numeric. Value of the soil property at depth3.}

\item{value4}{Numeric. Value of the soil property at depth4.}
}
\value{
Numeric. Weighted mean of soil values.
}
\description{
The purpose of this function is to calculate the weighted average of a soil
property across a soil column over standard depth intervals. This function
is designed to work on 4 depth breaks. This uses the trapezoidal rule as
described in Hengl et al. (2017, https://doi.org/10.1371/journal.pone.0169748).
}
