% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_data_pull.R, R/esd_esg_pull.R
\name{comp_data_pull}
\alias{comp_data_pull}
\title{Compile soil components for ecological site group of interest}
\usage{
comp_data_pull(
  file_paths_user = "Travis",
  target_ESG = "Semiarid_Warm_SandyUplands_LoamyUplands"
)

comp_data_pull(
  file_paths_user = "Travis",
  target_ESG = "Semiarid_Warm_SandyUplands_LoamyUplands"
)
}
\arguments{
\item{file_paths_user}{Character. User name to generate input data file paths.
Options are "Anna", "Travis", and "VPN".}

\item{target_ESG}{The ecological site group to pull data for.}
}
\value{
Soil Profile Collection


}
\description{
Compile soil components for ecological site group of interest

Pull ecological site data for the ecological site group of interest
}
