% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_species_lists.R
\name{compile_species_lists}
\alias{compile_species_lists}
\title{Compile species lists used to calculated grouped indicators}
\usage{
compile_species_lists(
  user = "Anna",
 
    output_path = "C:/Users/aknight/Desktop/Telework_Backups/V_drive/PROJECTS/ESG/Maps/Eco_Site_Group_Data"
)
}
\arguments{
\item{user}{Character. User name to generate input data file paths, passed to \link[STMdevelopment]{data_file_paths}}

\item{output_path}{Character. File path to the folder to saved the compiled species list in.}
}
\value{
Data frame containing the combined species list
}
\description{
Compile species lists used to calculated grouped indicators
}
