% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_file_paths.R
\name{data_file_paths}
\alias{data_file_paths}
\title{Set file paths for specific users' working environments}
\usage{
data_file_paths(user)
}
\arguments{
\item{user}{Character. Options are "Anna" and "Travis" for personal hard
drive and OneDrive folders/files, or VPN if doing everything from the
network}
}
\value{
List. Use this list in other functions to access data from different
file locations
}
\description{
The purpose of this function is to make it easy to run the analyses from
different users working environments. Add new list items as needed to access
other data sets.
}
