% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDIT_map.R
\name{EDIT_map}
\alias{EDIT_map}
\title{Create leaflet map of the target ESG}
\usage{
EDIT_map(target_ESG = "Semiarid_Warm_SandyUplands_LoamyUplands", user = "Anna")
}
\arguments{
\item{target_ESG}{Character. Ecological Site Group to be mapped.
Follow the names used in the ESG raster released with Travis's 2021 manuscript.}

\item{user}{Character. User name to generate input data file paths.
Options are "Anna", "Travis", and "VPN".}
}
\value{

}
\description{
Create leaflet map of the target ESG
}
