% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esg_production_pull.R
\name{esg_production_pull}
\alias{esg_production_pull}
\title{Ecological site group species production query}
\usage{
esg_production_pull(
  user = "Travis",
  target_ESG = "Semiarid_Warm_SandyUplands_LoamyUplands"
)
}
\arguments{
\item{user}{Character. User name to generate input data file paths.See
\code{\link{data_file_paths}} for input options.}

\item{target_ESG}{The ecological site group to pull data for.}
}
\value{
list of data.frames
}
\description{
Ecological site group species production query
}
