% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_allgaps.R
\name{estimate_allgaps}
\alias{estimate_allgaps}
\title{Estimate annual & perennial canopy gaps from perennial canopy gaps and LPI
cover data with a linear model built by Joe Brehm, to be published in McCord
et al. (in prep as of 10/2023). All estimates require values for (1) the
perennial gap window that matches the annual & perennial gap window to be
estimated (e.g. to estimate annual gaps > 25 cm, use perennial gaps > 25 cm),
(2) annual grass cover, (3) perennial grass cover, (4) shrub cover, and (5)
forb cover. All cover variables should come from any-hit indicators
calculated from LPI data. Structured to be used AFTER \link{plot_data_pull},
not with WITHIN \link{plot_data_pull}.}
\usage{
estimate_allgaps(gap_type = "25plus", perennial_gaps, AG, PG, SH, Forb)
}
\arguments{
\item{gap_type}{Character indicating the gap window to be predicted. One of
"25plus", "50plus", "100plus", or "200plus". Defaults to "25 plus".}

\item{perennial_gaps}{Numeric vector containing the perennial gap values.
Example indicator column name is "CP_percent_25plus".}

\item{AG}{Numeric vector containing the any-hit annual grass cover values.
Example indicator column name is "AH_AnnGrassCover".}

\item{PG}{Numeric vector containing the any-hit perennial grass cover values.
Example indicator column name is "AH_PerenGrassCover".}

\item{SH}{Numeric vector containing the any-hit shrub cover values. Example
indicator column name is "AH_ShrubCover".}

\item{Forb}{Numeric vector containing the any-hit forb cover values. Example
indicator column name is "AH_ForbCover".}
}
\value{
Numeric vector of estimated annual and perennial gap values for the
descriptor indicated in \code{gap_type}
}
\description{
Estimate annual & perennial canopy gaps from perennial canopy gaps and LPI
cover data with a linear model built by Joe Brehm, to be published in McCord
et al. (in prep as of 10/2023). All estimates require values for (1) the
perennial gap window that matches the annual & perennial gap window to be
estimated (e.g. to estimate annual gaps > 25 cm, use perennial gaps > 25 cm),
(2) annual grass cover, (3) perennial grass cover, (4) shrub cover, and (5)
forb cover. All cover variables should come from any-hit indicators
calculated from LPI data. Structured to be used AFTER \link{plot_data_pull},
not with WITHIN \link{plot_data_pull}.
}
