% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tables_EDIT_style.R
\name{format_tables_EDIT_style}
\alias{format_tables_EDIT_style}
\title{Format tables in RMarkdown to match EDIT website style}
\usage{
format_tables_EDIT_style(data, caption, row.names = F, col.names = NA)
}
\arguments{
\item{data}{Data frame to create table from}

\item{caption}{Character. Caption for table.}

\item{row.names}{Logical. Include row names in output table?}

\item{col.names}{Optional character vector of column names to be used in the table. Defaults to the column names of the data frame.}
}
\value{
Table formatted to match EDIT website
}
\description{
Format tables in RMarkdown to match EDIT website style
}
