% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gap100_predictive_model}
\alias{gap100_predictive_model}
\title{Predictive model for annual and perennial canopy gaps > 100 cm}
\format{
A linear model that predicts CA_percent_100plus (\% cover of annual and perennial canopy gaps > 100 cm)
\describe{
\item{CP_percent_100plus}{Independent variable representing \% cover of perennial canopy gaps > 100 cm}
\item{AH_AnnForbGrassCover}{Independent variable representing any-hit \% cover of annual herbaceous plants from LPI}
}
}
\usage{
gap100_predictive_model
}
\description{
This model can estimate percent cover of annual and perennial canopy gaps >
100 cm using percent cover of perennial canopy gaps > 100 cm and any-hit
annual herbaceous cover. It was built using data from all NRI and LMF plots
in the Upper Colorado River Basin where both perennial gaps and annual +
perennial gaps were measure at the same time. Adjusted R-squared:  0.8604
}
\keyword{datasets}
