% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dominant_spp.R
\name{get_dominant_spp}
\alias{get_dominant_spp}
\title{Get dominant species by functional group for each plant community}
\usage{
get_dominant_spp(
  data = filter(plot_data_descriptive, PlantCommunity_fuzzy == 3),
  species_cols = colnames(descriptive.df),
  user = "Anna"
)
}
\arguments{
\item{data}{Data frame containing all the species to be tallied. Format
should match the outputs of \code{\link{plot_data_pull}}. Filter the input
data frame to the plant community of interest to get dominant species by
community.}

\item{species_cols}{Character vector of the column names of all the columns
in data containing species cover.}

\item{user}{Character. User name to generate input data file paths. Options
are "Anna", "Travis", and "VPN". Passes to \code{\link{data_file_paths}}}
}
\value{
Data frame of species recorded for each functional group, arranged
from highest to lowest mean cover. Functional groups are:
C3NativePerenGrass, C3IntroducedPerenGrass, C4NativePerenGrass,
C4IntroducedPerenGrass, NativeAnnGrass, IntroducedAnnGrass,
NativePerenForb, IntroducedPerenForb, NativeAnnForb, IntroducedAnnForb,
Succulent, ShrubSubShrub, Tree
}
\description{
Get a list of the species in each functional group for each plant community,
ranked by mean cover across all plots in the plant community. Species must
appear in at least one plot to be returned.
}
