% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highrestiff.R
\name{highrestiff}
\alias{highrestiff}
\title{Save figure as a high-resolution TIFF file}
\usage{
highrestiff(plot_obj, file, width_in, height_in, resolution_dpi = 300)
}
\arguments{
\item{plot_obj}{Figure to be saved}

\item{file}{Character. Path and name of the file to save the figure in.}

\item{width_in}{Numeric. Width of the saved figure in inches}

\item{height_in}{Numeric. Height of the saved figure in inches}

\item{resolution_dpi}{Numeric. Image resolution in dots per square inch.}
}
\value{
List. Use this list in other functions to access data from different
file locations
}
\description{
This function saves a figure as a high resolution TIFF file
}
