% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_gaps.R
\name{impute_missing_gaps}
\alias{impute_missing_gaps}
\title{Impute missing annual & perennial canopy gaps with a linear model}
\usage{
impute_missing_gaps(
  indicator_data_target_wide,
  impute_gap_type = c("CA_percent_100plus", "CA_percent_200plus"),
  impute_sources = NULL
)
}
\arguments{
\item{indicator_data_target_wide}{Data frame. Intermediate data product within \link{plot_data_pull}}

\item{impute_gap_type}{Character vector. Gap type to predict. Supported options are
"CA_percent_100plus" (\% cover annual & perennial gaps >100 cm) and
"CA_percent_200plus" (\% cover annual & perennial gaps >200 cm)}

\item{impute_sources}{Character. Optional. Only impute missing data for
specific data source(s). Supported options are any source in
plot_data$SourceKey}
}
\value{
Wide format data frame containing all indicators for the target ESG,
including estimated annual & perennial canopy gaps for plots missing that
data type.
}
\description{
Impute missing annual & perennial canopy gaps with a linear model
}
