% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_indicator_descriptions.R
\name{make_indicator_descriptions}
\alias{make_indicator_descriptions}
\title{Create a table describing the plot indicators}
\usage{
make_indicator_descriptions(
  indicators = c("AH_C3PerenGrassCover", "AH_C4PerenGrassCover",
    "AH_IntroducedPerenGrassCover", "AH_NativePerenForbCover",
    "AH_IntroducedPerenForbCover", "AH_NativeAnnGrassCover",
    "AH_IntroducedAnnGrassCover", "AH_NativeAnnForbCover", "AH_IntroducedAnnForbCover",
    "BareSoilCover", "CP_percent_100to200", "CP_percent_200plus", "FH_LichenCover",
    "FH_MossCover"),
  shrub_by_spp = T,
  subshrub_by_spp = T,
  tree_by_spp = T,
  opuntia_combined = T
)
}
\arguments{
\item{indicators}{Character vector. Functional group indicators used to pull plot data}

\item{shrub_by_spp}{Logical. Include all shrubs by species?}

\item{subshrub_by_spp}{Logical. Include all sub-shrubs by species?}

\item{tree_by_spp}{Logical. Include all trees by species?}

\item{opuntia_combined}{Logical. Include combined cover of genus Opuntia?}
}
\value{
Data frame describing each indicator and set of indicators
}
\description{
Create a table describing the plot indicators
}
