% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NbClust_fanny.R
\name{NbClust_fanny}
\alias{NbClust_fanny}
\title{Calculate cluster validation metrics for FANNY method. See NbClust::NbClust() and cluster::fanny() for more details.}
\usage{
NbClust_fanny(
  data = NULL,
  diss = NULL,
  distance = "euclidean",
  min.nc = 2,
  max.nc = 15,
  method = NULL,
  index = "all",
  alphaBeale = 0.1,
  memb.exp = 2
)
}
\arguments{
\item{data}{matrix or dataset}

\item{diss}{dissimilarity matrix to be used. By default, diss=NULL, but if it is replaced by a dissimilarity matrix, distance should be "NULL".}

\item{distance}{the distance measure to be used to compute the dissimilarity matrix. This must be one of: "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski" or "NULL". By default, distance="euclidean". If the distance is "NULL", the dissimilarity matrix (diss) should be given by the user. If distance is not "NULL", the dissimilarity matrix should be "NULL".}

\item{min.nc}{minimal number of clusters, between 1 and (number of objects - 1)}

\item{max.nc}{maximal number of clusters, between 2 and (number of objects - 1), greater or equal to min.nc. By default, max.nc=15}

\item{method}{the cluster analysis method to be used. This should be one of: "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid", "kmeans"}

\item{index}{the index to be calculated. This should be one of : "kl", "ch", "hartigan", "ccc", "scott", "marriot", "trcovw", "tracew", "friedman", "rubin", "cindex", "db", "silhouette", "duda", "pseudot2", "beale", "ratkowsky", "ball", "ptbiserial", "gap", "frey", "mcclain", "gamma", "gplus", "tau", "dunn", "hubert", "sdindex", "dindex", "sdbw", "all" (all indices except GAP, Gamma, Gplus and Tau), "alllong" (all indices with Gap, Gamma, Gplus and Tau included).}

\item{alphaBeale}{significance value for Beale's index.}

\item{memb.exp}{value to pass to cluster::fanny()}
}
\value{
list containing index values, critical values, best number of clusters from each metric, and the best partitioning results
}
\description{
Calculate cluster validation metrics for FANNY method. See NbClust::NbClust() and cluster::fanny() for more details.
}
