% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMDS_scree.R
\name{NMDS_scree}
\alias{NMDS_scree}
\title{Create NMDS ordination scree plot to determine number of dimensions}
\usage{
NMDS_scree(x)
}
\arguments{
\item{x}{Ordination data frame containing only the indicator values (no plot identifiers)}
}
\value{
Plots stress versus number of dimensions in ordination solution
}
\description{
Create NMDS ordination scree plot to determine number of dimensions
}
