% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_mode.R
\name{raster_mode}
\alias{raster_mode}
\title{Calculate mode for aggregating ESG rasters to coarser resolution}
\usage{
raster_mode(x, na.rm = T)
}
\arguments{
\item{x}{Vector of values pulled from a raster. Can be numeric, character, factor, logical.}

\item{na.rm}{Logical. Omit NA values?}
}
\value{
Modal value of the inputs in same format as inputs. If multiple modes exist, returns the first appearing value of the set of modes.
}
\description{
Calculate mode for aggregating ESG rasters to coarser resolution
}
