% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_function.R
\name{template_function}
\alias{template_function}
\title{Bind two factors}
\usage{
template_function(a, b)
}
\arguments{
\item{a}{factor}

\item{b}{factor}
}
\value{
factor
}
\description{
Create a new factor from two existing factors, where the new factor's levels
are the union of the levels of the input factors.
}
\examples{
template_function(iris$Species[c(1, 51, 101)], PlantGrowth$group[c(1, 11, 21)])
}
