% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compiled_species_list}
\alias{compiled_species_list}
\title{Compiled species list from AIM, LMF, NRI, NCPN and USGS internal data}
\format{
\subsection{\code{compiled_species_list}}{

A data frame with 15320 rows and 14 columns:
\describe{
\item{SpeciesCode}{Character. Abbreviation unique to the species, typically based on the USDA Plants database (https://plants.usda.gov/home)}
\item{ScientificName}{ }
\item{CommonName}{ }
\item{Family}{ }
\item{Duration}{ }
\item{GrowthHabit}{ }
\item{GrowthHabitSub}{ }
\item{Native}{ }
\item{UpdatedSpeciesCode}{ }
\item{SpeciesState}{ }
\item{Noxious}{ }
\item{PhotosyntheticPathway}{ }
\item{SG_Group}{ }
\item{PinyonJuniper}{ }
}
}
}
\usage{
compiled_species_list
}
\description{
This file contains the species codes (typically derived from the USDA Plant
database), scientific names, and attributes of each species used for grouping
plants into functional types.
}
\keyword{datasets}
