% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esd_esg_pull.R
\name{esd_data_pull}
\alias{esd_data_pull}
\title{Pull ecological site data for the ecological site group of interest}
\usage{
esd_data_pull(
  user = "Travis",
  target_ESG = "Semiarid_Warm_SandyUplands_LoamyUplands"
)
}
\arguments{
\item{user}{Character. User name to generate input data file paths.
Options are "Anna", "Travis", and "VPN".}

\item{target_ESG}{The ecological site group to pull data for.}
}
\value{

}
\description{
Pull ecological site data for the ecological site group of interest
}
