% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esg_production_pull.R
\name{esg_production_pull}
\alias{esg_production_pull}
\title{Ecological site group species production query}
\usage{
esg_production_pull(
  file_paths_user = "Travis",
  target_ESG = "Semiarid_Warm_SandyUplands_LoamyUplands"
)
}
\arguments{
\item{file_paths_user}{Character. User name to generate input data file paths.
Options are "Anna", "Travis", and "VPN".}

\item{target_ESG}{The ecological site group to pull data for.}
}
\value{
list of data.frames
}
\description{
Ecological site group species production query
}
